Move code around to get rid of circular dependencies
